fix run.ps1

+++ dotnet/eShopOnWeb/scripts/run.ps1
@@ -1,8 +1,11 @@
 Set-Location -Path "C:\Users\Administrator\Desktop\eShopOnWeb\src\Web"
 $env:UseOnlyInMemoryDatabase = "true"
 $env:RedisServerUrl = "localhost:6379"
-$prev = Import-Clixml -Path (Join-Path 'C:\Users\Administrator' 'sandbox-process.xml')
-$prev | Stop-Process -ErrorAction SilentlyContinue
+$filePath = Join-Path 'C:\Users\Administrator' 'sandbox-process.xml'
+if (Test-Path -Path $filePath -PathType Leaf) {
+  $prev = Import-Clixml -Path $filePath
+  $prev | Stop-Process -ErrorAction SilentlyContinue
+}
 $process = Start-Process "dotnet" -ArgumentList "run --launch-profile Web" -PassThru
-$process | Export-Clixml -Path (Join-Path 'C:\Users\Administrator' 'sandbox-process.xml')
+$process | Export-Clixml -Path $filePath
 Wait-Process -Id $process.Id
